# board-service/docs/swagger.yaml
basePath: /api
definitions:
...
  project-board-api_internal_dto.CommentResponse:
    properties:
      boardId:
        type: string
      commentId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
...
  project-board-api_internal_dto.CreateCommentRequest:
    properties:
      boardId:
        type: string
      content:
        minLength: 1
        type: string
    required:
    - boardId
    - content
    type: object
...
  project-board-api_internal_dto.UpdateCommentRequest:
    properties:
      content:
        minLength: 1
        type: string
    required:
    - content
    type: object
...
  project-board-api_internal_response.ErrorResponse:
    properties:
      error: {}
      requestId:
        type: string
    type: object
  project-board-api_internal_response.SuccessResponse:
    properties:
      data: {}
      requestId:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    프로젝트 보드 관리 시스템 API 서버입니다.
    Board, Project, Comment, Participant 관리 기능을 제공합니다.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Project Board Management API
  version: "1.0"
paths:
...
  /comments:
    get:
      description: 특정 Board의 모든 Comment를 조회합니다. 프론트엔드 호환용 엔드포인트
      parameters:
      - description: Board ID (UUID)
        in: query
        name: boardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/project-board-api_internal_dto.CommentResponse'
                  type: array
              type: object
        "400":
          description: 잘못된 Board ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board의 Comment 목록 조회 (쿼리 파라미터 방식)
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Board에 새로운 Comment를 작성합니다
      parameters:
      - description: Comment 생성 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Comment 생성 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.CommentResponse'
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Comment 생성
      tags:
      - comments
  /comments/{commentId}:
    delete:
      description: Comment를 소프트 삭제합니다
      parameters:
      - description: Comment ID (UUID)
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment 삭제 성공
          schema:
            $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
        "400":
          description: 잘못된 Comment ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Comment를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Comment 삭제
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Comment 내용을 수정합니다
      parameters:
      - description: Comment ID (UUID)
        in: path
        name: commentId
        required: true
        type: string
      - description: Comment 수정 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Comment 수정 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.CommentResponse'
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Comment를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Comment 수정
      tags:
      - comments
  /comments/board/{boardId}:
    get:
      description: 특정 Board의 모든 Comment를 조회합니다
      parameters:
      - description: Board ID (UUID)
        in: path
        name: boardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/project-board-api_internal_dto.CommentResponse'
                  type: array
              type: object
        "400":
          description: 잘못된 Board ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board의 Comment 목록 조회
      tags:
      - comments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT 토큰을 입력하세요. 형식: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"